🐛 fix: 이미지 내에 괄호가 있을 경우 이미지 파싱이 안되는 문제 수정 및 파서 파일 core로 이동
