+++ apps/core/constants.py
@@ -1,0 +1,6 @@
+"""
+Core 앱 전역 상수 정의
+"""
+
+# 이미지 업로드 및 파싱에서 허용되는 확장자
+ALLOWED_IMAGE_EXTENSIONS = {".jpg", ".jpeg", ".png", ".gif"}
