+++ apps/posts/services/post_services.py
@@ -1,14 +1,14 @@
 from typing import Any, List, Optional
 
 from django.db import transaction
-from django.db.models import F, Q, QuerySet
+from django.db.models import F
 
 from apps.posts.exceptions.post_exceptions import (
     PostNotFoundException,
     PostPermissionDeniedException,
 )
 from apps.posts.models import Post, PostAttachment, PostImage
-from apps.qna.utils.content_parser import ContentParser
+from apps.core.utils.content_parser import ContentParser
 from apps.users.models import User
