+++ apps/qna/services/question/command.py
@@ -8,7 +8,7 @@ from rest_framework import status
 from apps.qna.constants import ErrorMessages
 from apps.qna.exceptions import QnaBaseException
 from apps.qna.models import Question, QuestionCategory, QuestionImage
-from apps.qna.utils.content_parser import ContentParser
+from apps.core.utils.content_parser import ContentParser
 from apps.qna.utils.model_types import User
 
 # ==============================================================================
